DTSCCI-683 Refactor ManageContactInformationCallbackHandler in civil-service (#5452)
* DTSCCI-683 Refactor ManageContactInformationCallbackHandler in civil-service - fix test

* DTSCCI-683 Refactor ManageContactInformationCallbackHandler in civil-service

* Add logging and fix open api test

* DTSCCI-683: Re-enable flyway for integration-tests

* Resolve conflict

---------
