+++ src/main/java/uk/gov/hmcts/reform/civil/model/UpdatePartyDetailsForm.java
@@ -1,10 +1,14 @@
 package uk.gov.hmcts.reform.civil.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder(toBuilder = true)
+@NoArgsConstructor
+@AllArgsConstructor
 public class UpdatePartyDetailsForm {
 
     private String firstName;
